--- generators/pulsar/files.mts
@@ -1,51 +1,0 @@
-/**
- * Copyright 2013-2023 the original author or authors from the JHipster project.
- *
- * This file is part of the JHipster project, see https://www.jhipster.tech/
- * for more information.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { WriteFileSection } from '../base/api.mjs';
-import { SERVER_TEST_SRC_DIR } from '../generator-constants.mjs';
-import { moveToJavaPackageTestDir } from '../server/support/index.mjs';
-import PulsarGenerator from './generator.mjs';
-
-export const pulsarFiles: WriteFileSection<any, any> = {
-  config: [
-    {
-      condition: data => data.buildToolGradle,
-      templates: ['gradle/pulsar.gradle'],
-    },
-  ],
-  test: [
-    {
-      path: `${SERVER_TEST_SRC_DIR}package/`,
-      renameTo: moveToJavaPackageTestDir,
-      templates: [
-        'broker/PulsarIT.java',
-        'config/BrokerConfiguration.java',
-        'config/EmbeddedPulsar.java',
-        'config/PulsarTestContainer.java',
-        'config/PulsarTestContainersSpringContextCustomizerFactory.java',
-      ],
-    },
-  ],
-};
-
-export default function writePulsarFilesTask(this: PulsarGenerator, { application }) {
-  return this.writeFiles({
-    sections: pulsarFiles,
-    context: application,
-  });
-}
