+++ generators/spring-cloud-stream/generator-pulsar.spec.mts
