+++ generators/spring-cloud-stream/templates/gradle/pulsar.gradle.ejs
