+++ generators/spring-cloud-stream/templates/src/test/java/package/broker/PulsarIT.java.ejs
