+++ generators/spring-cloud-stream/templates/src/test/java/package/config/BrokerConfiguration.java.ejs
