+++ generators/spring-cloud-stream/templates/src/test/java/package/config/EmbeddedPulsar.java.ejs
