+++ generators/spring-cloud-stream/templates/src/test/java/package/config/PulsarTestContainer.java.ejs
