+++ generators/spring-cloud-stream/templates/src/test/java/package/config/PulsarTestContainersSpringContextCustomizerFactory.java.ejs
